requirements.txt
Django==1.10.4
django-cors-headers==1.3.1
django-extensions==1.7.5
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
gdata==2.0.18
Pillow==3.4.2
requests==2.12.3
six==1.10.0
